greetBot-v2.0/requirements.txt
flask
RPi.GPIO
pygame
